+++ SpawnDev.VoxelEngine.Demo.Shared/UnitTests/VoxelEngineTestBase.Rendering.cs
@@ -1,7 +1,7 @@
 using System.Numerics;
 using System.Runtime.InteropServices;
 using ILGPU.Runtime;
-using SpawnDev.BlazorJS.JSObjects;
+using SpawnDev.SpawnJS.JSObjects;
 using SpawnDev.ILGPU.WebGPU;
 using SpawnDev.UnitTesting;
 using SpawnDev.VoxelEngine.Meshing;